rcs: moved 'Move_Command' in extra file
